Refactor: Removed erraneous packaging, I know it's convention to do the whole com.company, etc, but it's ugz
