+++ src/bot/App.java
@@ -1,4 +1,4 @@
-package com.jboby93.markovbot;
+package bot;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
