+++ src/bot/MarkovBot.java
@@ -1,4 +1,4 @@
-package com.jboby93.markovbot;
+package bot;
 
 import java.io.IOException;
